fix: Extract scrollToSection util and fix hero CTA scroll + dead index prop
